"multitouch" in BossLevel enabled, Shoot timer changed , started working on Boss-Player-Collision

+++ JosieGame/Classes/PlayerBoss.h
@@ -12,7 +12,7 @@ public:
 
 	void moveLeft();
 	void moveRight();
-	void shoot(int counterForShoot);
+	void shoot(float counterForShoot);
 	void jump();
 
 	void useShot(int id);
